+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 0.0.10
     - Add support for override content language 
     - Add support for the fulfillment API
+    - Add support for override marketplace 
     - Add support for the payment policy API
     - Add support for the return policy API
+    - Add support for the searchByImage API
 
 0.0.9
   - Add support for trading APIs, shopping APIs & Finding APIs